mzcompose: split out a services module
This commit is pure code movement that splits out a services module in
mzcompose. This separates the definition of builtin services from the
innards of mzcompose, which lowers the barrier to entry for adding a new
service or modifying an existing service.
