+++ test/cluster/mzworkflows.py
@@ -7,13 +7,13 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-from materialize.mzcompose import (
+from materialize.mzcompose import Workflow
+from materialize.mzcompose.services import (
     Coordd,
     Dataflowd,
     Kafka,
     SchemaRegistry,
     Testdrive,
-    Workflow,
     Zookeeper,
 )
