+++ test/feature-benchmark/mzworkflows.py
@@ -24,12 +24,12 @@ from materialize.feature_benchmark.termination import (
     NormalDistributionOverlap,
     ProbForMin,
 )
-from materialize.mzcompose import (
+from materialize.mzcompose import Workflow
+from materialize.mzcompose.services import (
     Kafka,
     Materialized,
     SchemaRegistry,
     Testdrive,
-    Workflow,
     Zookeeper,
 )
