+++ test/limits/mzworkflows.py
@@ -13,12 +13,12 @@ import os
 import sys
 import tempfile
 
-from materialize.mzcompose import (
+from materialize.mzcompose import Workflow
+from materialize.mzcompose.services import (
     Kafka,
     Materialized,
     SchemaRegistry,
     Testdrive,
-    Workflow,
     Zookeeper,
 )
