+++ test/proxy/mzworkflows.py
@@ -7,14 +7,14 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-from materialize.mzcompose import (
+from materialize.mzcompose import Workflow
+from materialize.mzcompose.services import (
     Kafka,
     Localstack,
     Materialized,
     SchemaRegistry,
     Squid,
     Testdrive,
-    Workflow,
     Zookeeper,
 )
