+++ test/sqllogictest/mzworkflows.py
@@ -7,7 +7,8 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-from materialize.mzcompose import Postgres, SqlLogicTest, Workflow
+from materialize.mzcompose import Workflow
+from materialize.mzcompose.services import Postgres, SqlLogicTest
 
 services = [Postgres(), SqlLogicTest()]
